- Add queries folder - Ad PageDetail page - wip Profil Detail integration
