+++ app/components/Profil/ProfilCard.tsx
@@ -15,7 +15,7 @@ interface IProps {
 
 export default function ProfilCard({ profil }: IProps) {
     return (
-        <Link href="/profil">
+        <Link href={`/profil/${profil.id}`}>
             <div className="flex flex-col rounded-lg overflow-hidden shadow-profil-card shadow-shadow p-4 gap-4">
                 <div><Image className="rounded-md" width={318} height={477} style={{ objectFit: 'cover', width: '100%' }} src={profil.photo} alt="image profil" /></div>
                 <p className="text-lg font-semibold">{profil.name}, {profil.age}</p>
